[feat]ロジックをcomposableに分離
